Implement upload book image and Add new validations
Now is possible to upload image for a book. The gems 'carrierwave' and
'mini_magick' is responsible for doing that task. So they are add to the
Gemfile, an image uploader class were generated and also an initializer
for carrier_wave.

Livro model class is now validating the following:
 - Uniqueness for 'nome', 'autor' and 'editora' are now require before
add a new book
 - The fields 'nome', 'autor' and 'editora' can't be blank
 - The field 'nota' must be a integer and < 6
 - The uploaded images should be less than 500KB